cr fixes: rename layout components
