+++ src/components/layouts/AuthPageLayout.vue
