+++ src/components/layouts/WithSidebarLayout.vue
